Upsert method - inserts new or updates existing element
Upsert method uses callback which return value is inserted
into a map. Callback is given 3 arguments:
  - exists - whether given key already exists in map
  - valueInMap - existing value in map
  - valueBeingAdded - value which was passed to Upsert method

Return value of a callback is used to save new value in map.

This approach allows map users to do non-trivial transformations
of value stored in map, while still protected by a lock

It is somewhat ovelaps with #29, but more generic and allows
more use cases