fix `Keys` function returns `nil` keys when concurrently removing.
Now the `Keys` function return a fixed length key slice whose length
is calculated at the beginning of the function. When the `map` removes
element during the function body, the returned result will contain `nil`
keys, in our case, empty strings. This commit fixes this bug by using
`append` function on a zero length slice with the initial length of the
map as `cap`.