Fix deadlock when using Iter/IterBuffered
As mentioned by @efeller. When Using the current implementation,
after calling Iter/IterBuffered, if the returned channel were
not drained, some readlock may not be properly released. Which
cause subsequent call to change the map, namely Set/Remove/Pop
operations to block forever, thus a deadlock occurs.

We introduce two test functions (TestUnDrainedIter and
TestUnDrainedIterBuffered) which demonstrate the bug.

This main idea of the fix is to take a snapshot of the current
concurrentMap when calling Iter/IterBuffered. We factor out a
`snapshot` function here, which returns an array of buffered
channels(one per each shard, the readlocks are released after the
related channel is populated). They then fan in to a return channel.
In the buffered case, the size of the returned channel is the sum
of length of all the channels. We avoid using `m.Count()` here since
`m` may change between the function calls, which makes it inaccurate.

This fix passes those two test functions mentioned before.